atualização do projeto

+++ ServerSideSD/resources/views/users/all_users.blade.php
@@ -12,20 +12,33 @@
     <h1>Aqui vês todos os users</h1>
 
     <hr><hr>
-    <p>Name: {{ $cesaeInfo['name']}}</p>
-    <p>Address: {{ $cesaeInfo['address']}}</p>
-    <p>Email: {{ $cesaeInfo['email']}}</p>
-    <hr><hr><br><br>
-
-
+    <p>Name: {{ $contactPerson->name}}</p>
+    <p>Address: {{ $contactPerson->address}}</p>
+    <p>Email: {{ $contactPerson->email}}</p>
     <hr><hr>
+
+    {{-- pesquisa --}}
+
+    <div class="container-fluid text-center">
+
+        <form action="">
+            <input type="text" id="search" name="search" value="{{ request()->query('search') }}">
+            <button type="submit" class="btn btn-secondary">Procurar</button>
+        </form>
+
+    </div>
+
+
+
+
+
+    <hr><hr><br>
     <table class="table table-dark table-striped table-hover text-center">
 
         <tr>
             <td>ID</td>
             <td>Name</td>
             <td>Email</td>
-            <td></td>
             <td></td>
         </tr>
 
@@ -36,14 +49,19 @@
                     <td>{{ $array->name}}</td>
                     <td>{{ $array->email}}</td>
                     <td><a class="btn btn-info" href="{{ route('users.showone', $array->id) }}">Ver</a>
-                    <a class="btn btn-danger" href="{{ route('users.deleteOne', $array->id) }}">Apagar</a>
-                    <a class="btn btn-danger" href="{{ route('users.update', $array->id) }}">Modificar</a></td>
+                    <a class="btn btn-danger" href="{{ route('users.update', $array->id) }}">Modificar</a>
+                        @auth
+                            @if (Auth::user()->email == 'admin@gmail.com')
+                                <a class="btn btn-danger" href="{{ route('users.deleteOne', $array->id) }}">Apagar</a>
+                            @endif
+                        @endauth
+                    </td>
                 </tr>
 
         @endforeach
 
     </table>
-    <hr><hr>
+    <br><hr><hr><br>
 
 
 @endsection
